+++ src/_multithread.js
@@ -3,7 +3,7 @@ const cluster = require("node:cluster");
 if (cluster.isMaster) {
     const electron = require("electron");
     const ipc = electron.ipcMain;
-    const signale = require("signale");
+    const signale = require("./utils/logger.js");
     // Also, leave a core available for the renderer process
     const osCPUs = require("node:os").cpus().length - 1;
     // See #904
@@ -72,7 +72,7 @@ if (cluster.isMaster) {
         }
     });
 } else if (cluster.isWorker) {
-    const signale = require("signale");
+    const signale = require("./utils/logger.js");
     const si = require("systeminformation");
 
     signale.info("Multithread worker started at "+process.pid);